rpmsg: glink: Remove IRQF_NO_SUSPEND.
Currently interrupts with hardware wake capabilities still cause the
CPUs to handle the interrupt at the GIC layer when IRQF_NO_SUSPEND is
used. This is reproduced in both s2idle and mem suspend.

Remove IRQF_NO_SUSPEND and add enable_irq_wake to properly account for
the wakeup interrupt.

Change-Id: Ieb8de40102fb402ca2f040020f56b8e7c42b3611
Bug: 123377615
Bug: 131260677
Signed-off-by: fadlyas07 <mhmmdfdlyas@gmail.com>