UPSTREAM: psi: Fix cpu.pressure for cpu.max and competing cgroups
For simplicity, cpu pressure is defined as having more than one
runnable task on a given CPU. This works on the system-level, but it
has limitations in a cgrouped reality: When cpu.max is in use, it
doesn't capture the time in which a task is not executing on the CPU
due to throttling. Likewise, it doesn't capture the time in which a
competing cgroup is occupying the CPU - meaning it only reflects
cgroup-internal competitive pressure, not outside pressure.

Enable tracking of currently executing tasks, and then change the
definition of cpu pressure in a cgroup from

	NR_RUNNING > 1

to

	NR_RUNNING > ON_CPU

which will capture the effects of cpu.max as well as competition from
outside the cgroup.

After this patch, a cgroup running `stress -c 1` with a cpu.max
setting of 5000 10000 shows ~50% continuous CPU pressure.

Signed-off-by: Johannes Weiner <hannes@cmpxchg.org>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Link: https://lkml.kernel.org/r/20200316191333.115523-2-hannes@cmpxchg.org
(cherry picked from commit b05e75d611380881e73edc58a20fd8c6bb71720b)

BUG=b:244260899
TEST=CQ

Signed-off-by: Zubin Mithra <zsm@google.com>
Change-Id: I7a25fa641fb9035019ab79e10f17535972a5e23b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3872018
Tested-by: Zubin Mithra <zsm@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Zubin Mithra <zsm@chromium.org>
Signed-off-by: fadlyas07 <mhmmdfdlyas@gmail.com>