sched: Resolve sched_feat() at compile time to improve code optimization
Keeping the compiler in the dark about which sched_feat() toggles are
set reduces its ability to optimize scheduler code, which results in a
measurably significant performance loss since the affected code is very
hot. Turning sched_feat() definitions into macros that the compiler can
resolve results in smaller and faster code.

The sched_feat() definitions are converted to macros with this sed:
sed -i -E \
	-e 's/SCHED_FEAT\((.*),(.*)\)/#define SCHED_FEAT_\1\2/' \
	-e 's/true/1/' \
	-e 's/false/0/' \
	kernel/sched/features.h

The sched_feat() macros are named SCHED_FEAT_*. With this done, we can
now optimize try-to-wake-up code when SCHED_FEAT_TTWU_QUEUE is disabled,
which results in a CPU usage reduction of a few percent inside
try_to_wake_up().

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[0ctobot: Adapt for 4.19]
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Change-Id: I11e6c4bc49e12454f1e977b7d761c5ba07c7a0f0
Signed-off-by: fadlyas07 <mhmmdfdlyas@gmail.com>