FIXUP: devfreq_boost: Introduce devfreq boost driver
This driver boosts enumerated devfreq devices upon input, and allows for
boosting specific devfreq devices on other custom events. The boost
frequencies for this driver should be set so that frame drops are
near-zero at the boosted frequencies and power consumption is minimized
at said frequencies. The goal of this driver is to provide an interface
to achieve optimal device performance by requesting boosts on key
events, such as when a frame is ready to rendered to the display.

Currently, support is only present for boosting the cpubw devfreq
device, but the driver is structured in a way that makes it easy to add
support for new boostable devfreq devices in the future.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[0ctobot: Bring inline with kerneltoast/android_kernel_google_floral@dce5aa5,
amending the following commits where applicable:
076de0a ("[2/2] devfreq_boost: Rework display state awareness logic for OnePlus 8 devices")
bf07438 ("devfreq_boost: Update and expand to handle CPUBW/LLCCBW boosting")
2ca6410 ("devfreq_boost: Mark boost kthreads as performance critical")]
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Change-Id: Iffda967c6a61a5e5d98f49bbd2be780b1cf4b705
Signed-off-by: fadlyas07 <mhmmdfdlyas@gmail.com>