ARM64/configs: juice: Don't randomize module region over a 4 GB range
Randomizing a module is for security purpose to deter exploit
attempts relying on knowledge of the location of kernel internals.

But randomizing the location of module region inside a 4 GB window
to prevent core kernel data leakage is expensive. We are better off
with a limited range. Randomizing module region is expensive anyway.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: fadlyas07 <mhmmdfdlyas@gmail.com>