work-simple: Rewrite for clarity and performance
This implementation suffers from a race on init where a timer interrupt may
occur before `glob_worker` is set, causing the kernel to panic as a result.
Upon closer inspection, all of the kthread juggling inside the swork code
appears to make no sense, and the get/put semantics are also meaningless
considering that there are core kernel users that do a swork_get() on init
and then never do a swork_put(). A spin lock is also used for a linked list
that could simply be a lock-less list instead.

Rewrite the swork code to remove the get/put semantics and speed things up
with a lock-less list. Now this code truly is simple.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: fadlyas07 <mhmmdfdlyas@proton.me>