soc: qcom: watchdog_v2: Adapt the watchdog to RT
Firstly, make the watchdog IRQ non-threaded on RT to ensure that the bark
handler runs at all and in a timely manner in case the scheduler is dead.

Secondly, perform a normal panic() rather than a watchdog bite on RT so
that the printk log buffers get flushed when there's a watchdog bark. This
is necessary on RT because the spin locks in printk() can sleep, so it's
only possible to flush the log buffers from an IRQ when oops_in_progress is
set. Otherwise, the ramoops gets truncated when the system reboots due to a
forced watchdog bite.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: fadlyas07 <mhmmdfdlyas@proton.me>