rcu: Don't force rcu_normal on RT for Android
Android works much better with rcu_expedited, since it speeds up a lot of
procedures that are important to Android such as process death, which is
needed to free up processes' resources quickly when memory runs low.

Since Android isn't designed to be a real-time userspace anyway, remove the
rcu_normal restrictions on RT for Android.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: fadlyas07 <mhmmdfdlyas@proton.me>