mm: Fix truncated major/minor output in PID maps
The major and minor can be up to 12 bits and 20 bits long, respectively.
The current output generated after the micro-optimizations applied to
show_vma_header_prefix() only show up to 8 bits of each (2 hex digits),
resulting in the major and minor being truncated.

Fix it by introducing new optimized macros to print up to 3 and 5 hex
digits for the major and minor, respectively.

Reported-by: LoveSy <shana@zju.edu.cn>
Reported-by: 南宫雪珊 <vvb2060@gmail.com>
Change-Id: I0973c6f7b1d8c83f07c4bc41f5ed77fc5b2bb840
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: fadlyas07 <mhmmdfdlyas@proton.me>