ARM64/configs: vendor: Merge fast-forward from main defconfig to bengal-perf_defconfig
This source works on two devices with one defconfig, converting to fragment defconfig seems like a waste of time as there won't be any conflicts to worry about in the future. Also, copying and creating a new defconfig from arch defconfig would be a waste of time as we would have to push every change manually during CLO tag updates.

Change the stock defconfig prefix because we don't use that codename anymore.

Signed-off-by: fadlyas07 <mhmmdfdlyas@proton.me>